+++ src/NetAppFiles/NetAppFiles/Models/PSNetAppFilesCache.cs
@@ -87,5 +87,11 @@ namespace Microsoft.Azure.Commands.NetAppFiles.Models
         public string GlobalFileLocking { get; set; }
 
         public string WriteBack { get; set; }
+
+        /// <summary>
+        /// Gets flag indicating whether file access logs are enabled for the Cache,
+        /// based on active diagnostic settings present on the Cache. Possible values include: &#39;Enabled&#39;, &#39;Disabled&#39;
+        /// </summary>
+        public string FileAccessLogs { get; set; }
     }
 }